Implemented a basic logging method
After checking the arguments for the 'debug' keyword the logger either prints info or doesn't

the listen function of the networking module has to be changed to not accept the logger as a parameter
but rather initialize a listener class with the logger.
